Updated SettingsFile Resource
Added Setters for all variables to ensure we are emitting when we have changed.
Organized settings into regions for better readability.
Added bunch of Setttings.
Added snapshot() and restore(), to take a snapshot of the current settings, so if the user modifies values, but cancels the dialog, we restore the settings back to their original values.