Created LabelTimer
Created a Label Timer, that allows to to create a countdown, count up,
and fuzzy countup of time.
When countdown is set, count's down from the start_count, and when
hitting 0, will display countdown_complete_text.
When countdown is false, it counts up from start till it is stopped.
When fuzzy time is setup, will display the seconds till minutes, minutes
till hours, and hours till days.