Updated Twitcher Plugin
Added couple of patches to fix some issues.
TwitchAuth:  force_verify was not properly bubbled down to OAuth to ensure that we are actually fully requesting the renew of chat tokens.
TwitchService:  Checked to see if we are Authenticated or not.  Only call auth.authorize() if we are not Authorized yet.